PlayerSuperCards Renamed
Deleted old PlayerCards component and renamed PlayerSuperCards to PlayerCards in every instance.
