+++ ComposureWatch/src/Components/Card/CardList.jsx
@@ -1,7 +1,6 @@
 import React from "react";
 
 import PlayerCards from "./PlayerCards";
-import PlayerSuperCards from "./PlayerSuperCards";
 
 const CardList = (props) => {
   if (props.cards.length === 0) {
@@ -11,7 +10,7 @@ const CardList = (props) => {
   return (
     <ul className="grid col-1 sm:grid-cols-3 md:grid-cols-6 gap-x-6 gap-y-8">
       {props.cards.map((card) => (
-        <PlayerSuperCards
+        <PlayerCards
           id={card.id}
           name={card.name}
           character={card.character}
